Use Dynamic Text as a plugin
This commit demonstrates how to use dynamic text in a component.
All entry points must use the DynamicTextPlugin.

Some parameters that come from the environment are not done yet.

The `getDynamicCampaignText` should be refactored that
a) each value is only created once
b) each value is only created when requested