Update form styles and messages
Added the styles for custom amount and upgrade forms.

Added language items for custom amount and upgrade forms.

Made all sub forms into forms again. This is because it
gives us and easier way to handle keyboard navigation when
we do the accessibility pass.

Updated the provided currency formatter as multiple methods
are needed.

Renamed the input formatter to euro formatter as it is used
for more than inputs.

Added tests for checking that the correct values are inserted
into the copy.