Change close event tracking
Renamed userInteraction to userChoice because that words makes it more
clear that it's a choice between several options and not the name of an
interaction.

Have a unified close event that doesn't use CloseSources, but translates
them to feature and userChoice pairs. The switch statement is
unfortunate, maybe we can find a better mechanism.

Adapt legacy event maps

Co-Authored-By: Gabriel Birke <gabriel.birke@wikimedia.de>
Co-Authored-By: Corinna Hillebrand <corinna.hillebrand@wikimedia.de>
Co-Authored-By: Sperling-0 <tanuja.doriya@wikimedia.de>