Refactor close tracking event
The close events are now created in the components
and passed up the tree.

The CloseSources enum is renamed to LegacyCloseSources
and marked as deprecated.

The long switch statement is removed from the close event
and the conversion is done in the legacy closeMapper.

Updated the WPDE banner mappings to track a single close
event.