Improve mapping of NotShownEvent
This maps all NotShownEvent instances that don't have "SizeIssue" as a
reason to a WMDELegacyBannerEvent with a trackingRate of 0 (we don't
need to track those ATM).

Add a test for mapNotShownEvent since the map now contains conditionals.

NotShownEvent now adds a default reason.

NotShownState now always adds the reason when tracking the
NotShownEvent.

Delete mapSizeIssueEvent because it duplicated mapNotShownEvent and was
not used anywhere.