Change CTRL banner for precampaign
 The first slide is removed from both banners.
The feature of redirecting users directly to the payment providers is removed in both banners.
The soft-close feature is removed/not used in both banners.
The progress bar is removed from the modal/expanded banner in both banners.
The campaign day sentence is removed/not used in both banners.
The visitors-vs-donors sentence is removed/replaced in both banners.
broken styles are fixed (wikipedia skin introduced some weird borders)
The banners share the same height with the banner they are based on.

https://phabricator.wikimedia.org/T366161