Convert DE Desktop to new swatches
I realised that if that the more banners we have when
they swap the default skin to Vector 2022 that the more
work it'll be to swap all the old ones over to it.

I did the effort work in now to swap our existing ones over
and if they do change the skin during the campaign we will
have saved ourselves some headaches.