Add dark mode swatch to EN Mobile
Enwiki Minerva now supports dark mode so we need to add a
dark swatch to the theme for the en mobile banner.

This also add swatching to the payment provider icons, which
is a backwards breaking change and may need to be integrated
into existing banners after being merged.

Ticket: https://phabricator.wikimedia.org/T376794