Change evaluation of is-late-progress
We evaluate if the campaigns are in late progress by
looking at a html data attribute that we store in a
mediawiki template included with our banners.

If this attribute contains any string it was being
evaluated as true. This change makes it check that
the attribute value is explicitly the string `true`.