Move current campaigns tests into folders
Get a baseline to check if the new vitest configuration works (two kinds
of tests in desktop) and to provide a structure for other tests.

Renamed the wikipedia.de channels in `campaign_info.toml` to match
folder names to channel names. This now must be adjusted in
https://github.com/wmde/banner-screenshots/blob/main/src/CommandLine/determineCampaignFromBranchName.ts

Turn missing tests into errors.