Fix Treedip theme
For the 2024 Desktop WPDE 01 banner I introduced the `$transaction-fees`
parameter to the swatches. I placed it in the middle of the parameter
list because I thought it semantically belonged to the "select-group"
features. This accidentally shifted the positional parameters of
skin_vector-2022 because it did not contain the new parameter.

Since the parameter `$transaction-fees` was ultimately not used (we
opted for using a general checkbox style in the donation form
components), this commit removes the parameter again to fix the English
banners which use Treedip with `skin_vector-2022`.

The `skin_default.scss` and `skin_vector-2022.scss` are turning the
keyword parameters into positional parameters. We might look into
turning them into keyword parameters again in the future.