Change local development page to wiki/Hamster
The front pages of the wikis contain sections for
news. This means that sometimes there's pictures
of notable terrible people on there which the
developers then have to see all day while working
on a banner.

This changes the development page to wiki/Hamster
so we can look at some cute fur bundles instead.