Implement VAR for C24_WMDE_Mobile_DE_11
We wanted to optimize the markup for the close button, but ran into
issues where the test code depended on the class `wmde-banner-mini-close-button` being present.
We added back the class to make the tests pass, but will optimize our
tests and markup. Tracked in #645

Ticket: https://phabricator.wikimedia.org/T380854