Fix soft close button of C24_WMDE_Desktop_DE_17
The close button in the upper right corner was not hooked up to an event
handler. After checking the code and trying out the banner, I concluded
that overriding the close button is not necessary in this instance.

The test that checks the close function of the soft close only started
to fail after the branch was merged.

Ticket: https://phabricator.wikimedia.org/T380464