Add Vue ESLint rules
Copied the rules from the [mediawiki coding style](https://github.com/wikimedia/eslint-config-wikimedia)

Marked & commented out rules that will change the code significantly,
those will need a discussion & decision with the whole team.

Commented out some rules that'll be part of the commit that
automatically fixes violations.

Fixed/changed some components that had violations. The components in
DonationForm/Forms all have an `isCurrent` property that's no longer
used in the markup. I decided to ignore the notification for now,
because the banners component still pass `is-current` to the forms and
the property could be useful in the future.