Create domain classes for form actions
We need to modify the form actions dynamically during A/B tests and want
to encapsulate the URL generation logic inside an API.

This commit is backwards-compatible with the existing `FormActions`
interface, but deprecates it.

Since the impression counting is non-reactive but tries to be reactive,
we don't put it as a dependency of `FormAction`. When it becomes
reactive, we should change the constructor of `FormAction` and hand over
the impression counting class as a dependency.

Ticket: https://phabricator.wikimedia.org/T383980