Hide banner when user is temporary
When a user makes an edit, they get a temporary user ID, but do not
count as a "logged in" user for CentralNotice.

The banner for notifying the user about their temporary ID obscures the
fundraising banner, so we hide it in these cases.

Ticket: https://phabricator.wikimedia.org/T408908